[ISSUE #359] Split handler from controller (#359)
